+++ codex-rs/core/src/exec_policy_tests.rs
@@ -34,6 +34,10 @@ use tempfile::TempDir;
 use tempfile::tempdir;
 use toml::Value as TomlValue;
 
+#[cfg(windows)]
+#[path = "exec_policy_windows_tests.rs"]
+mod windows_tests;
+
 fn config_stack_for_dot_codex_folder(dot_codex_folder: &Path) -> ConfigLayerStack {
     let dot_codex_folder =
         AbsolutePathBuf::from_absolute_path(dot_codex_folder).expect("absolute dot_codex_folder");
@@ -656,64 +660,17 @@ async fn evaluates_bash_lc_inner_commands() {
     .await;
 }
 
-#[cfg(windows)]
-#[tokio::test]
-async fn evaluates_powershell_inner_commands_against_prompt_rules() {
-    assert_exec_approval_requirement_for_command(
-        ExecApprovalRequirementScenario {
-            policy_src: Some(r#"prefix_rule(pattern=["echo"], decision="prompt")"#.to_string()),
-            command: vec![
-                "powershell.exe".to_string(),
-                "-NoProfile".to_string(),
-                "-Command".to_string(),
-                "echo blocked".to_string(),
-            ],
-            approval_policy: AskForApproval::Never,
-            sandbox_policy: SandboxPolicy::DangerFullAccess,
-            file_system_sandbox_policy: unrestricted_file_system_sandbox_policy(),
-            sandbox_permissions: SandboxPermissions::UseDefault,
-            prefix_rule: None,
-        },
-        ExecApprovalRequirement::Forbidden {
-            reason: PROMPT_CONFLICT_REASON.to_string(),
-        },
-    )
-    .await;
-}
-
-#[cfg(windows)]
-#[tokio::test]
-async fn evaluates_powershell_inner_commands_against_allow_rules() {
-    assert_exec_approval_requirement_for_command(
-        ExecApprovalRequirementScenario {
-            policy_src: Some(r#"prefix_rule(pattern=["echo"], decision="allow")"#.to_string()),
-            command: vec![
-                "powershell.exe".to_string(),
-                "-NoProfile".to_string(),
-                "-Command".to_string(),
-                "echo blocked".to_string(),
-            ],
-            approval_policy: AskForApproval::UnlessTrusted,
-            sandbox_policy: SandboxPolicy::new_read_only_policy(),
-            file_system_sandbox_policy: read_only_file_system_sandbox_policy(),
-            sandbox_permissions: SandboxPermissions::UseDefault,
-            prefix_rule: None,
-        },
-        ExecApprovalRequirement::Skip {
-            bypass_sandbox: true,
-            proposed_execpolicy_amendment: None,
-        },
-    )
-    .await;
-}
-
 #[test]
 fn commands_for_exec_policy_falls_back_for_empty_shell_script() {
     let command = vec!["bash".to_string(), "-lc".to_string(), "".to_string()];
 
     assert_eq!(
         commands_for_exec_policy(&command),
-        (vec![command], false, ExecPolicyCommandOrigin::Direct)
+        ExecPolicyCommands {
+            commands: vec![command],
+            used_complex_parsing: false,
+            command_origin: ExecPolicyCommandOrigin::Generic,
+        }
     );
 }
 
@@ -727,84 +684,12 @@ fn commands_for_exec_policy_falls_back_for_whitespace_shell_script() {
 
     assert_eq!(
         commands_for_exec_policy(&command),
-        (vec![command], false, ExecPolicyCommandOrigin::Direct)
+        ExecPolicyCommands {
+            commands: vec![command],
+            used_complex_parsing: false,
+            command_origin: ExecPolicyCommandOrigin::Generic,
+        }
     );
-}
-
-#[cfg(windows)]
-#[test]
-fn commands_for_exec_policy_parses_powershell_shell_wrapper() {
-    let command = vec![
-        "powershell.exe".to_string(),
-        "-NoProfile".to_string(),
-        "-Command".to_string(),
-        "echo blocked".to_string(),
-    ];
-
-    assert_eq!(
-        commands_for_exec_policy(&command),
-        (
-            vec![vec!["echo".to_string(), "blocked".to_string()]],
-            false,
-            ExecPolicyCommandOrigin::PowerShell,
-        )
-    );
-}
-
-#[cfg(windows)]
-#[test]
-fn unmatched_safe_powershell_words_are_allowed() {
-    let command = vec!["Get-Content".to_string(), "Cargo.toml".to_string()];
-
-    assert_eq!(
-        Decision::Allow,
-        render_decision_for_unmatched_command(
-            &command,
-            UnmatchedCommandContext {
-                approval_policy: AskForApproval::UnlessTrusted,
-                permission_profile: &permission_profile_from_sandbox_policy(
-                    &SandboxPolicy::new_read_only_policy(),
-                ),
-                file_system_sandbox_policy: &read_only_file_system_sandbox_policy(),
-                sandbox_cwd: Path::new("/tmp"),
-                sandbox_permissions: SandboxPermissions::UseDefault,
-                used_complex_parsing: false,
-                command_origin: ExecPolicyCommandOrigin::PowerShell,
-            },
-        )
-    );
-}
-
-#[cfg(windows)]
-#[tokio::test]
-async fn unmatched_dangerous_powershell_inner_commands_require_approval() {
-    let inner_command = vec![
-        "Remove-Item".to_string(),
-        "test".to_string(),
-        "-Force".to_string(),
-    ];
-
-    assert_exec_approval_requirement_for_command(
-        ExecApprovalRequirementScenario {
-            policy_src: None,
-            command: vec![
-                "powershell.exe".to_string(),
-                "-NoProfile".to_string(),
-                "-Command".to_string(),
-                "Remove-Item test -Force".to_string(),
-            ],
-            approval_policy: AskForApproval::OnRequest,
-            sandbox_policy: SandboxPolicy::DangerFullAccess,
-            file_system_sandbox_policy: unrestricted_file_system_sandbox_policy(),
-            sandbox_permissions: SandboxPermissions::UseDefault,
-            prefix_rule: None,
-        },
-        ExecApprovalRequirement::NeedsApproval {
-            reason: None,
-            proposed_execpolicy_amendment: Some(ExecPolicyAmendment::new(inner_command)),
-        },
-    )
-    .await;
 }
 
 #[tokio::test]
@@ -1110,7 +995,7 @@ fn unmatched_granular_policy_still_prompts_for_restricted_sandbox_escalation() {
                 sandbox_cwd: Path::new("/tmp"),
                 sandbox_permissions: SandboxPermissions::RequireEscalated,
                 used_complex_parsing: false,
-                command_origin: ExecPolicyCommandOrigin::Direct,
+                command_origin: ExecPolicyCommandOrigin::Generic,
             },
         )
     );
@@ -1132,7 +1017,7 @@ fn unmatched_on_request_uses_split_filesystem_policy_for_escalation_prompts() {
                 sandbox_cwd: Path::new("/tmp"),
                 sandbox_permissions: SandboxPermissions::RequireEscalated,
                 used_complex_parsing: false,
-                command_origin: ExecPolicyCommandOrigin::Direct,
+                command_origin: ExecPolicyCommandOrigin::Generic,
             },
         )
     );
@@ -2117,10 +2002,20 @@ struct ExecApprovalRequirementScenario {
     prefix_rule: Option<Vec<String>>,
 }
 
-async fn assert_exec_approval_requirement_for_command(
+fn policy_from_src(policy_src: Option<&str>) -> Arc<Policy> {
+    match policy_src {
+        Some(src) => {
+            let mut parser = PolicyParser::new();
+            parser.parse("test.rules", src).expect("parse policy");
+            Arc::new(parser.build())
+        }
+        None => Arc::new(Policy::empty()),
+    }
+}
+
+async fn exec_approval_requirement_for_command(
     test: ExecApprovalRequirementScenario,
-    expected_requirement: ExecApprovalRequirement,
-) {
+) -> ExecApprovalRequirement {
     let ExecApprovalRequirementScenario {
         policy_src,
         command,
@@ -2131,19 +2026,10 @@ async fn assert_exec_approval_requirement_for_command(
         prefix_rule,
     } = test;
 
-    let policy = match policy_src {
-        Some(src) => {
-            let mut parser = PolicyParser::new();
-            parser
-                .parse("test.rules", src.as_str())
-                .expect("parse policy");
-            Arc::new(parser.build())
-        }
-        None => Arc::new(Policy::empty()),
-    };
+    let policy = policy_from_src(policy_src.as_deref());
 
     let permission_profile = permission_profile_from_sandbox_policy(&sandbox_policy);
-    let requirement = ExecPolicyManager::new(policy)
+    ExecPolicyManager::new(policy)
         .create_exec_approval_requirement_for_command(ExecApprovalRequest {
             command: &command,
             approval_policy,
@@ -2153,8 +2039,14 @@ async fn assert_exec_approval_requirement_for_command(
             sandbox_permissions,
             prefix_rule,
         })
-        .await;
+        .await
+}
 
+async fn assert_exec_approval_requirement_for_command(
+    test: ExecApprovalRequirementScenario,
+    expected_requirement: ExecApprovalRequirement,
+) {
+    let requirement = exec_approval_requirement_for_command(test).await;
     assert_eq!(requirement, expected_requirement);
 }
