+++ codex-rs/shell-command/src/powershell.rs
@@ -74,9 +74,11 @@ pub fn extract_powershell_command(command: &[String]) -> Option<(&str, &str)> {
 /// This is intentionally narrower than the Windows safe-command parser: it only unwraps the
 /// `-Command`/`-c` body from a PowerShell invocation we already recognize, then delegates the
 /// script itself to the PowerShell AST parser.
-pub fn parse_powershell_command_plain_commands(command: &[String]) -> Option<Vec<Vec<String>>> {
-    let (shell, script) = extract_powershell_command(command)?;
-    try_parse_powershell_ast_commands(shell, script)
+pub fn parse_powershell_command_into_plain_commands(
+    command: &[String],
+) -> Option<Vec<Vec<String>>> {
+    let (executable, script) = extract_powershell_command(command)?;
+    try_parse_powershell_ast_commands(executable, script)
 }
 
 /// This function attempts to find a powershell.exe executable on the system.
@@ -151,7 +153,7 @@ fn is_powershellish_executable_available(powershell_or_pwsh_exe: &std::path::Pat
 mod tests {
     use super::extract_powershell_command;
     #[cfg(windows)]
-    use super::parse_powershell_command_plain_commands;
+    use super::parse_powershell_command_into_plain_commands;
 
     #[test]
     fn extracts_basic_powershell_command() {
@@ -203,7 +205,7 @@ mod tests {
     #[cfg(windows)]
     #[test]
     fn parses_plain_powershell_commands() {
-        let commands = parse_powershell_command_plain_commands(&[
+        let commands = parse_powershell_command_into_plain_commands(&[
             "powershell.exe".to_string(),
             "-NoProfile".to_string(),
             "-Command".to_string(),
@@ -212,5 +214,25 @@ mod tests {
         .expect("parse");
 
         assert_eq!(commands, vec![vec!["echo".to_string(), "hi".to_string()]]);
+    }
+
+    #[cfg(windows)]
+    #[test]
+    fn parses_multiple_plain_powershell_commands() {
+        let commands = parse_powershell_command_into_plain_commands(&[
+            "powershell.exe".to_string(),
+            "-NoProfile".to_string(),
+            "-Command".to_string(),
+            "Write-Output foo | Measure-Object".to_string(),
+        ])
+        .expect("parse");
+
+        assert_eq!(
+            commands,
+            vec![
+                vec!["Write-Output".to_string(), "foo".to_string()],
+                vec!["Measure-Object".to_string()],
+            ]
+        );
     }
 }
